Add files via upload
Raw data from experiment. Multiple tabs in the excel workbook are different forms of the data. The first tab is the primary data used in the analysis. The second tab is a metadata sheet explaining variable in the primary data. The third tab is the raw observational data recorded during the experiment. The other three tabs are supplementary data and summary data collected during the experiment but not used in the analysis.